snprintf: moved self-test to a separated test
