+++ .gitignore
@@ -49,6 +49,7 @@ tests/test_rand
 tests/test_sasl
 tests/test_scram
 tests/test_sha1
+tests/test_snprintf
 tests/test_sock
 m4/
 libstrophe.project
